+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -331,6 +331,56 @@ namespace Libplanet.Tests.Blockchain
         }
 
         [Fact]
+        public void Render()
+        {
+            DumbAction.RenderRecords.Value = ImmutableList<RenderRecord>.Empty;
+            MinerReward.RenderRecords.Value = ImmutableList<RenderRecord>.Empty;
+
+            var policy = new BlockPolicy<DumbAction>();
+            var fx = new DefaultStoreFixture(memory: true);
+            var key = new PrivateKey();
+            var miner = key.ToAddress();
+
+            try
+            {
+                var chain = new BlockChain<DumbAction>(
+                    policy, fx.Store, fx.GenesisBlock, render: false);
+                var actions = new[] { new DumbAction(miner, "foo") };
+                var tx = chain.MakeTransaction(key, actions);
+                Assert.Empty(DumbAction.RenderRecords.Value);
+
+                var block = TestUtils.MineNext(
+                    chain.Genesis,
+                    new[] { tx },
+                    miner: miner,
+                    difficulty: _blockChain.Policy.GetNextBlockDifficulty(_blockChain));
+                Assert.Empty(DumbAction.RenderRecords.Value);
+
+                // Render should not work when appending a block
+                chain.Append(block);
+                Assert.Empty(DumbAction.RenderRecords.Value);
+
+                // Render should not work when appending a block to forked chain
+                var forked = chain.Fork(chain.Genesis.Hash);
+                forked.Append(block);
+                Assert.Empty(DumbAction.RenderRecords.Value);
+
+                // Render should not work when swapping the chain
+                var newChain = new BlockChain<DumbAction>(
+                    policy, fx.Store, Guid.NewGuid(), fx.GenesisBlock, render: true);
+                chain.Swap(newChain, true);
+                chain.Append(block);
+                Assert.Empty(DumbAction.RenderRecords.Value);
+            }
+            finally
+            {
+                DumbAction.RenderRecords.Value = ImmutableList<RenderRecord>.Empty;
+                MinerReward.RenderRecords.Value = ImmutableList<RenderRecord>.Empty;
+                fx.Dispose();
+            }
+        }
+
+        [Fact]
         public void Append()
         {
             DumbAction.RenderRecords.Value = ImmutableList<RenderRecord>.Empty;
@@ -1896,7 +1946,8 @@ namespace Libplanet.Tests.Blockchain
                 new NullPolicy<DumbAction>(),
                 store,
                 chainId,
-                _fx.GenesisBlock
+                _fx.GenesisBlock,
+                true
             );
             var privateKey = new PrivateKey();
             Address signer = privateKey.ToAddress();